macos: first-smoke bring-up -- compat shims, build fixes, format sweep, #44 fix
Squash of iterative macOS-port bring-up work on top of the clean 7-commit
plan (HOST_DARWIN probe → posix_shims → IO_BACKEND_DARWIN gate → kqueue
substrate extraction → backend_darwin.c → scripts/docs).  This commit
captures everything from the first reffsd build through the end-to-end
read+write smoke from a Linux NFS client.

Darwin libc / SunRPC compatibility shims
----------------------------------------
* lib/include/reffs/darwin_rpc_compat.h (new)
    - xdr_sizeof() counting-stream shim: Darwin's base-libc SunRPC
      lacks xdr_sizeof(3).  Stack-allocated struct xdr_ops whose
      x_putlong adds 4 and x_putbytes adds len (no pad rounding -- pad
      bytes arrive via a second x_putbytes call per xdr_opaque).  Unit
      test pending (see issue #25).
* lib/include/reffs/task.h
    - On __APPLE__, pre-#include <mach/task.h> then #define task_resume
      reffs_task_resume / task_suspend reffs_task_suspend.  Avoids
      collision with Mach kernel APIs that get transitively included.
* lib/rpc/rpc.c
    - rpc_call_xdr() dispatcher: on Darwin, xdrproc_t is 3-arg
      (int (*)(XDR *, void *, unsigned int)); pass depth=0.  Elsewhere
      2-arg as before.
* lib/utils/server.c, lib/nfs4/client/ec_io.c
    - clock_nanosleep(CLOCK_MONOTONIC, 0, &ts, NULL) -> nanosleep(&ts,
      NULL) on Darwin (lacks clock_nanosleep, equivalent for relative
      sleeps).
* lib/nfs4/server/copy.c
    - Static-inline stub for copy_file_range(2) on Darwin + if (0) wrap
      around the zero-copy branch, forcing the existing read+write
      fallback.  Byte-range zero-copy on Darwin tracked as issue #21.

configure.ac Darwin branch
--------------------------
* CPPFLAGS aliases: -Dxdr_uint32_t=xdr_u_int32_t / xdr_uint64_t,
  -DRPCSEC_GSS=6, -DAUTH_SYS=AUTH_UNIX, -DAUTH_DH=3,
  -Dauthsys_create=authunix_create, -Dst_{atim,mtim,ctim}=st_*spec,
  -DUUID_STR_LEN=37.  Tracked for cleanup as issue #24.
* -include rpc/rpc.h to pull <rpc/types.h> (bool_t) for generated XDR.
* -Wl,-rpath,/opt/homebrew/lib so make check binaries find
  libhdr_histogram et al. without DYLD_LIBRARY_PATH.
* -Wl,-undefined,dynamic_lookup for flat-namespace late binding; matches
  Linux's default ld behavior.  Masks missing LIBADDs -- audit tracked
  as issue #48.
* AC_CHECK_DECLS (header-based) replacing AC_CHECK_FUNCS (link-based)
  for accept4 / pipe2.  dynamic_lookup otherwise makes the link probe
  lie.
* AX_CHECK_COMPILE_FLAG probe for -Wno-cast-function-type-mismatch
  (Apple CLT clang 20+ / upstream 20+) with -Wno-cast-function-type-
  strict fallback (upstream clang 19 and earlier, Homebrew llvm@18).
* Warning suppressions for clang -Wextra additions: -Wno-sign-compare,
  -Wno-pointer-sign, -Wno-incompatible-pointer-types-discards-
  qualifiers, -Wno-deprecated-non-prototype, and -Wno-variadic-macros
  / -Wno-c23-extensions / -Wno-gnu for liburcu headers.  TODOs tracked
  in issues #20 and #24 for proper follow-up.
* PKG_CHECK_MODULES / AX_PTHREAD call sites reshaped for autoconf 2.73
  strict-mode compat (AS_IF + full [] quoting).

Format-string sweep
-------------------
* ~127 substitutions across ~20 files: %lu -> PRIu64, %ld -> PRId64,
  %lx -> PRIx64 for uint64_t/int64_t args.  #include <inttypes.h> in
  each touched file.  off_t -> (intmax_t)+%jd, dev_t -> (unsigned long)
  +%lu.  -Werror=format re-enabled with no warnings.

Build system
------------
* lib/nfs4/server/Makefile.am, lib/nfs3/Makefile.am,
  lib/probe1/Makefile.am: LIBADD libreffs_dstore.la.  Required on
  macOS -- noinst static convenience archives don't auto-fold into
  dylibs otherwise; dstore_ops_local would be unresolved at dyld time.
* lib/Makefile.am + lib/nfs4/Makefile.am: SUBDIRS reorder so
  nfs4/dstore builds before consumers (nfs3, nfs4/server, probe1).
* lib/fs/tests/Makefile.am: gate fuse_1..fuse_9 off on HOST_DARWIN
  (reffs/fuse.h can't prototype without macFUSE headers).  Gate
  fs_usage_posix off too -- uses Linux CLONE_NEWUSER / CLONE_NEWNS.

Darwin-specific test & smoke assets
-----------------------------------
* examples/reffsd-macos-smoke.toml -- port 3049 standalone config,
  POSIX backend against /Volumes/reffs_data.
* docs/macos-install.md -- install runbook: Homebrew deps, bootstrap
  script, ACLOCAL_PATH quirk, known gotchas.
* docs/freebsd-{install,smoke}.md + .claude/patterns/backend-
  extraction.md: add SPDX headers (were missing, tripped the nightly
  license check).
* scripts/macos-bootstrap.sh: one-shot from-source install for
  liburcu, HdrHistogram_c, and xdr-parser (via pipx from loghyr/reply
  repo -- PyPI reply-xdr entry is macOS-unreliable).

Bug fix -- issue #44
--------------------
* lib/io/backend_darwin.c submit_file_op: take rpc_trans_get(rt) when
  stashing rt in ic->ic_rt, matching Linux (backend.c:278,311) and
  FreeBSD (backend_kqueue.c:298,336).  Without it the completion
  handler's rpc_protocol_free(rt) drops a ref the backend never
  acquired, freeing rt while the resumed task is still using it -- the
  NULL deref at rpc.c:667 that SIGSEGV'd reffsd on first WRITE.

End-to-end validation
---------------------
* macOS 14 arm64 (Apple CLT clang 21 + Homebrew llvm@18): build clean,
  make check green.
* NFSv4.1 mount from Linux client (Rocky 10.1 x86_64, garbo):
  mount + ls + getattr + readdir works.
* 16 MB dd + sha256 round-trip via NFS WRITE path: matches; reffsd
  survives.

Review follow-ups tracked as issues #45 (pipe backpressure missed
wakeup), #46 (worker fault-path RCU safety), #47 (pool->running
atomicity), #48 (LIBADD audit under -Wl,-undefined,error).  Issue #44
(write-path SIGSEGV) closed by this commit.

+++ lib/fs/fs.c
@@ -8,6 +8,7 @@
 #endif
 
 #include <errno.h>
+#include <inttypes.h>
 #include <limits.h>
 #include <stdatomic.h>
 #include <pthread.h>
@@ -339,7 +340,7 @@ int reffs_fs_create(const char *path, mode_t mode)
 	}
 
 	trace_fs_inode(dir, __func__, __LINE__);
-	TRACE("ino=%lu freeing struct at %p", dir->i_ino, (void *)dir);
+	TRACE("ino=%" PRIu64 " freeing struct at %p", dir->i_ino, (void *)dir);
 
 	ret = vfs_create(dir, nm->nm_name, mode, &ap, NULL, NULL, NULL);
 	inode_active_put(dir);
@@ -419,7 +420,7 @@ int reffs_fs_mknod(const char *path, mode_t mode, dev_t rdev)
 	int ret;
 	struct authunix_parms ap;
 
-	TRACE("path=%s mode=0%o rdev=%lu", path, mode, rdev);
+	TRACE("path=%s mode=0%o rdev=%lu", path, mode, (unsigned long)rdev);
 
 	ret = find_matching_directory_entry(&nm, path, LAST_COMPONENT_IS_NEW);
 	if (ret)
@@ -468,7 +469,8 @@ int reffs_fs_symlink(const char *target, const char *linkpath)
 
 int reffs_fs_fallocate(const char *path, int mode, off_t offset, off_t len)
 {
-	TRACE("path=%s mode=0%o offset=%lu len=%lu", path, mode, offset, len);
+	TRACE("path=%s mode=0%o offset=%jd len=%jd", path, mode,
+	      (intmax_t)offset, (intmax_t)len);
 	return 0;
 }
 
@@ -563,7 +565,7 @@ int reffs_fs_read(const char *path, char *buffer, size_t size, off_t offset)
 
 	int ret;
 
-	TRACE("path=%s size=%lu offset=%lu", path, size, offset);
+	TRACE("path=%s size=%lu offset=%jd", path, size, (intmax_t)offset);
 
 	ret = find_matching_directory_entry(&nm, path, LAST_COMPONENT_IS_MATCH);
 	if (ret)
@@ -610,7 +612,7 @@ out:
 int reffs_fs_readdir(const char *path, void *buffer, char *filler, off_t offset)
 {
 	int ret = 0;
-	TRACE("path=%s offset=%lu", path, offset);
+	TRACE("path=%s offset=%jd", path, (intmax_t)offset);
 
 	TRACE("ret=%d", ret);
 	return ret;
@@ -844,7 +846,7 @@ int reffs_fs_write(const char *path, const char *buffer, size_t size,
 
 	int ret;
 
-	TRACE("path=%s size=%lu offset=%lu", path, size, offset);
+	TRACE("path=%s size=%lu offset=%jd", path, size, (intmax_t)offset);
 
 	ret = find_matching_directory_entry(&nm, path, LAST_COMPONENT_IS_MATCH);
 	if (ret)
@@ -924,12 +926,12 @@ void reffs_fs_recover(struct super_block *sb)
 	if (!sb || !sb->sb_ops || !sb->sb_ops->recover)
 		return;
 
-	TRACE("Starting recovery for sb %lu from %s", sb->sb_id,
+	TRACE("Starting recovery for sb %" PRIu64 " from %s", sb->sb_id,
 	      sb->sb_backend_path ? sb->sb_backend_path : "(none)");
 
 	sb->sb_ops->recover(sb);
 
-	TRACE("Recovery complete. Max inode: %lu", sb->sb_next_ino);
+	TRACE("Recovery complete. Max inode: %" PRIu64, sb->sb_next_ino);
 }
 
 int reffs_fs_usage(struct reffs_fs_usage_stats *stats)
@@ -962,7 +964,9 @@ int reffs_fs_usage(struct reffs_fs_usage_stats *stats)
 	else
 		stats->free_files = 0;
 
-	TRACE("internal: used_bytes=%lu free_bytes=%lu total_bytes=%lu used_files=%lu free_files=%lu total_files=%lu",
+	TRACE("internal: used_bytes=%" PRIu64 " free_bytes=%" PRIu64
+	      " total_bytes=%" PRIu64 " used_files=%" PRIu64
+	      " free_files=%" PRIu64 " total_files=%" PRIu64,
 	      stats->used_bytes, stats->free_bytes, stats->total_bytes,
 	      stats->used_files, stats->free_files, stats->total_files);
